refactor(chat-direct.component.ts): Moved chat direct to the chat directory and changed it into a standalone component
