+++ src/BurstChat.Client/src/app/components/chat/chat-direct/chat-direct.component.html
@@ -1,0 +1,24 @@
+@if(chatFound) {
+    <div class="h-100">
+        <burst-chat-info [options]="options"/>
+        @switch(layoutState) {
+            @case('chat') {
+                <div class="direct-chat-content">
+                    <burst-chat-messages [options]="options"/>
+                </div>
+                <burst-chat-input [options]="options"/>
+            }
+            @case('call') {
+                <div class="h-100">
+                    <burst-chat-call [options]="options"/>
+                </div>
+            }
+        }
+    </div>
+    <burst-message-edit-dialog [(visible)]="editMessageData.visible" [options]="options"
+        [message]="editMessageData.message"/>
+    <burst-message-delete-dialog [(visible)]="deleteMessageData.visible" [options]="options"
+        [message]="deleteMessageData.message"/>
+} @else {
+    <div></div>
+}
