+++ src/BurstChat.Client/src/app/components/chat/chat-direct/chat-direct.component.scss
@@ -1,5 +1,5 @@
-@import '../../../../../scss/theme';
-@import '../../../../../scss/sizing';
+@import '../../../../scss/theme';
+@import '../../../../scss/sizing';
 
 .direct-chat-content {
     @extend .h-100;
