+++ src/BurstChat.Client/src/app/modules/chat/chat.routing.ts
@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { ChatRootComponent } from 'src/app/modules/chat/components/chat-root/chat-root.component';
-import { ChatDirectComponent } from 'src/app/modules/chat/components/chat-direct/chat-direct.component';
 import { ChatChannelComponent } from 'src/app/components/chat/chat-channel/chat-channel.component';
+import { ChatDirectComponent } from 'src/app/components/chat/chat-direct/chat-direct.component';
 
 const routes: Routes = [
     {
