--- src/BurstChat.Client/src/app/modules/chat/components/chat-direct/chat-direct.component.html
@@ -1,27 +1,0 @@
-<div class="h-100" *ngIf="!noChatFound">
-
-    <burst-chat-info [options]="options"></burst-chat-info>
-
-    <div class="direct-chat-content" *ngIf="layoutState === 'chat'">
-
-        <burst-chat-messages [options]="options"></burst-chat-messages>
-
-    </div>
-
-    <burst-chat-input [options]="options" *ngIf="layoutState === 'chat'"></burst-chat-input>
-
-    <div class="h-100" *ngIf="layoutState === 'call'">
-
-        <burst-chat-call [options]="options"></burst-chat-call>
-
-    </div>
-
-</div>
-
-<burst-message-edit-dialog [(visible)]="editMessageData.visible" [options]="options"
-    [message]="editMessageData.message"></burst-message-edit-dialog>
-
-<burst-message-delete-dialog [(visible)]="deleteMessageData.visible" [options]="options"
-    [message]="deleteMessageData.message"></burst-message-delete-dialog>
-
-<div *ngIf="noChatFound"></div>
